- Refactor and move the escrow mock service and mock data as per the identical idea as used every where. - Write the escrow mock service test cases using mock service later change the mock service with actual whenever market mock service is merged in market.
